# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "College API Documentation",
        "description": "API sučelje za upravljanje studentima, odjelima i profesorima u sustavu fakulteta.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Lokalni razvojni server"
        }
    ],
    "paths": {
        "/api/departments": {
            "get": {
                "tags": [
                    "Odjeli"
                ],
                "summary": "Dohvati listu odjela",
                "description": "Vraća paginiranu listu svih odjela.",
                "operationId": "getDepartmentsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Broj stranice za paginaciju",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspješno dohvaćanje odjela",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Računarstvo"
                                                    },
                                                    "hod_id": {
                                                        "type": "integer",
                                                        "example": 1,
                                                        "nullable": true
                                                    },
                                                    "head_of_department": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "first_name": {
                                                                "type": "string",
                                                                "example": "Pero"
                                                            },
                                                            "last_name": {
                                                                "type": "string",
                                                                "example": "Perić"
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup / Potrebna autentifikacija.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Odjeli"
                ],
                "summary": "Kreiraj novi odjel",
                "operationId": "storeDepartment",
                "requestBody": {
                    "description": "Podaci za novi odjel",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 30,
                                        "example": "Novi Odjel"
                                    },
                                    "hod_id": {
                                        "description": "ID postojećeg profesora (faculty) kao HOD-a",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Odjel uspješno kreiran",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Novi Odjel"
                                        },
                                        "hod_id": {
                                            "type": "integer",
                                            "example": 1,
                                            "nullable": true
                                        },
                                        "head_of_department": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Pero"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Perić"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Greška validacije",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Neautoriziran pristup / Potrebna autentifikacija.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic",
                "description": "HTTP Basic Authentication. Koristite email kao korisničko ime i odgovarajuću lozinku."
            }
        }
    },
...
}